Delete, Update

+++ app/src/main/java/chanh/androidlab4/MainActivity.java
@@ -11,6 +11,7 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ListView;
+import android.widget.Toast;
 
 import java.util.ArrayList;
 
@@ -24,8 +25,13 @@ public class MainActivity extends AppCompatActivity {
 
     // layout object
     EditText mEtName;
+    EditText mEtID;
     Button mBtInsert;
     Button mBtRead;
+    Button mBtDelete;
+    Button mBtUpdate;
+    Button mBtReset;
+    Button mBtSelect;
 
     ListView mList;
     ArrayAdapter<String> baseAdapter;
@@ -37,24 +43,32 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         // // Database 생성 및 열기
-        db = openOrCreateDatabase(dbName,dbMode,null);
+        db = openOrCreateDatabase(dbName, dbMode, null);
         // 테이블 생성
         createTable();
 
         mEtName = (EditText) findViewById(R.id.et_text);
+        mEtID = (EditText) findViewById(R.id.et_Id);
         mBtInsert = (Button) findViewById(R.id.bt_insert);
         mBtRead = (Button) findViewById(R.id.bt_read);
+        mBtDelete = (Button) findViewById(R.id.bt_delete);
+        mBtUpdate = (Button) findViewById(R.id.bt_update);
+        mBtReset = (Button) findViewById(R.id.bt_reset);
+        mBtSelect = (Button) findViewById(R.id.bt_select);
         ListView mList = (ListView) findViewById(R.id.list_view);
 
         mBtInsert.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String name = mEtName.getText().toString();
-                insertData(name);
+                if (mEtName.getText().toString().equals(""))
+                    Toast.makeText(getApplicationContext(), "TEXT 입력하세요", Toast.LENGTH_SHORT).show();
+                else {
+                    String name = mEtName.getText().toString();
+                    insertData(name);
+                }
             }
         });
 
-        mBtRead = (Button) findViewById(R.id.bt_read);
         mBtRead.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -66,9 +80,57 @@ public class MainActivity extends AppCompatActivity {
 
         // Create listview
         nameList = new ArrayList<String>();
-        baseAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_dropdown_item_1line, nameList);
+        baseAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_expandable_list_item_1, nameList);
         mList.setAdapter(baseAdapter);
 
+        mBtDelete.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (mEtID.getText().toString().equals(""))
+                    Toast.makeText(getApplicationContext(), "ID 입력하세요", Toast.LENGTH_SHORT).show();
+                else {
+                    String id = mEtID.getText().toString();
+                    removeData(Integer.parseInt(id));
+                }
+            }
+        });
+
+        mBtUpdate.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (mEtID.getText().toString().equals("") && mEtName.getText().toString().equals(""))
+                    Toast.makeText(getApplicationContext(), "다시 입력하세요", Toast.LENGTH_SHORT).show();
+                else if (mEtName.getText().toString().equals(""))
+                    Toast.makeText(getApplicationContext(), "TEXT 입력하세요", Toast.LENGTH_SHORT).show();
+                else if (mEtID.getText().toString().equals(""))
+                    Toast.makeText(getApplicationContext(), "ID 입력하세요", Toast.LENGTH_SHORT).show();
+                else {
+                    String id = mEtID.getText().toString();
+                    String name = mEtName.getText().toString();
+                    updateData(Integer.parseInt(id), name);
+                }
+            }
+        });
+
+        mBtReset.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                removeTable();
+                createTable();
+            }
+        });
+
+        mBtSelect.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (mEtID.getText().toString().equals(""))
+                    Toast.makeText(getApplicationContext(), "ID 입력하세요", Toast.LENGTH_SHORT).show();
+                else {
+                    String id = mEtID.getText().toString();
+                    selectData(Integer.parseInt(id));
+                }
+            }
+        });
     }
 
     // Table 생성
@@ -77,7 +139,7 @@ public class MainActivity extends AppCompatActivity {
             String sql = "create table " + tableName + "(id integer primary key autoincrement, " + "name text not null)";
             db.execSQL(sql);
         } catch (android.database.sqlite.SQLiteException e) {
-            Log.d("Lab sqlite","error: "+ e);
+            Log.d("Lab sqlite", "error: " + e);
         }
     }
 
@@ -101,8 +163,16 @@ public class MainActivity extends AppCompatActivity {
 
     // Data 삭제
     public void removeData(int index) {
-        String sql = "delete from " + tableName + " where id = " + index + ";";
-        db.execSQL(sql);
+        String sql = "select * from " + tableName + " where id = " + index + ";";
+        Cursor result = db.rawQuery(sql, null);
+
+        if (result.moveToFirst()) {
+            sql = "delete from " + tableName + " where id = " + index + ";";
+            db.execSQL(sql);
+            Toast.makeText(this, "삭제했습니다", Toast.LENGTH_SHORT).show();
+        } else {
+            Toast.makeText(this, "비어있습니다", Toast.LENGTH_SHORT).show();
+        }
     }
 
     // Data 읽기(꺼내오기)
@@ -114,8 +184,10 @@ public class MainActivity extends AppCompatActivity {
         if (result.moveToFirst()) {
             int id = result.getInt(0);
             String name = result.getString(1);
-//            Toast.makeText(this, "index= " + id + " name=" + name, Toast.LENGTH_LONG).show();
+            Toast.makeText(this, "index= " + id + " name=" + name, Toast.LENGTH_SHORT).show();
             Log.d("lab_sqlite", "\"index= \" + id + \" name=\" + name ");
+        } else {
+            Toast.makeText(this, "비어있습니다", Toast.LENGTH_SHORT).show();
         }
         result.close();
     }
@@ -129,7 +201,7 @@ public class MainActivity extends AppCompatActivity {
         while (!results.isAfterLast()) {
             int id = results.getInt(0);
             String name = results.getString(1);
-//            Toast.makeText(this, "index= " + id + " name=" + name, Toast.LENGTH_LONG).show();
+//            Toast.makeText(this, "index= " + id + " name=" + name, Toast.LENGTH_SHORT).show();
             Log.d("lab_sqlite", "index= " + id + " name=" + name);
 
             nameList.add(name);
@@ -137,5 +209,4 @@ public class MainActivity extends AppCompatActivity {
         }
         results.close();
     }
-
 }
